pyproject.toml
[project]
name = "symfony-php-mcp"
version = "0.1.0"
description = "MCP server bridging Symfony/PHP projects with LLMs via Claude Desktop and other MCP clients"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["mcp", "symfony", "php", "llm", "ai", "claude"]
authors = [{ name = "symfony-php-mcp contributors" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "mcp[cli]>=1.0.0",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/maschmann/symfony-php-mcp"
Repository = "https://github.com/maschmann/symfony-php-mcp"
Issues = "https://github.com/maschmann/symfony-php-mcp/issues"

[project.scripts]
symfony-mcp = "symfony_mcp.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/symfony_mcp"]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
